bullhead: wifi: enable and tune BET
Enabling Beacon Early Termination in order
to speed up beacon processing drastically.
This will lead to power savings.

Also tune the BET wake interval to a value
of "8" to reduce the frequency of wakeups.
"8" is a safe choice since I could detect values
of higher than "10" to cause issues in some usecases.

Change-Id: I943f01b7599d8f7e42b1bc9249b8fc9b783e4345
Signed-off-by: Alex Naidis <alex.naidis@linux.com>